# api/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Bored Manager API
  version: 0.1.0
  summary: Versioned control-plane API for Bored Manager.
  description: |
    Public REST and event contracts for Bored Manager. All mutating operations
    require both the session cookie and an `X-CSRF-Token` header. Resource IDs
    are canonical lowercase UUIDv7 values. Timestamps are RFC 3339 UTC values.
    Event cursors are decimal strings so JavaScript clients do not lose integer
    precision. This development contract may only change compatibly within v1.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://127.0.0.1:8443/api/v1
    description: Initial local setup endpoint
  - url: https://{manager}/api/v1
    variables:
      manager:
        default: bored-manager.local:8443
security:
  - sessionCookie: []
...
paths:
...
  /setup:
    get:
      tags: [Setup]
      operationId: getSetupStatus
      security: []
      responses:
        '200':
          description: Current first-run setup state.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/SetupStatus'}
    post:
      tags: [Setup]
      operationId: completeSetup
      description: Available only over loopback while setup is incomplete.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/CompleteSetupRequest'}
      responses:
        '201':
          description: Setup completed and initial admin session created.
          headers:
            Set-Cookie: {$ref: './components.yaml#/components/headers/SessionCookie'}
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/Session'}
        '400': {$ref: './components.yaml#/components/responses/BadRequest'}
        '403': {$ref: './components.yaml#/components/responses/Forbidden'}
        '409': {$ref: './components.yaml#/components/responses/Conflict'}
        '422': {$ref: './components.yaml#/components/responses/ValidationFailed'}
        '429': {$ref: './components.yaml#/components/responses/RateLimited'}
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Authenticated session.
          headers:
            Set-Cookie: {$ref: './components.yaml#/components/headers/SessionCookie'}
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/Session'}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
        '429': {$ref: './components.yaml#/components/responses/RateLimited'}
  /auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      responses:
        '200':
          description: Current session.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/Session'}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
    delete:
      tags: [Auth]
      operationId: logout
      parameters: [{$ref: './components.yaml#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Session revoked.}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
        '403': {$ref: './components.yaml#/components/responses/Forbidden'}
  /auth/reauthenticate:
    post:
      tags: [Auth]
      operationId: reauthenticate
      parameters: [{$ref: './components.yaml#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/ReauthenticateRequest'}
      responses:
        '200':
          description: Session marked recently reauthenticated.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/Session'}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
        '403': {$ref: './components.yaml#/components/responses/Forbidden'}
        '429': {$ref: './components.yaml#/components/responses/RateLimited'}
  /agents:
    get:
      tags: [Agents]
      operationId: listAgents
      parameters:
        - {$ref: './components.yaml#/components/parameters/PageSize'}
        - {$ref: './components.yaml#/components/parameters/PageToken'}
        - {name: presence, in: query, schema: {$ref: './components.yaml#/components/schemas/PresenceState'}}
        - {name: tag, in: query, schema: {type: string}}
        - {name: search, in: query, schema: {type: string, maxLength: 128}}
      responses:
        '200':
          description: Cursor-consistent agent snapshot.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/AgentList'}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
  /agents/install-command:
    get:
      tags: [Agents]
      operationId: getAgentInstallCommand
      description: Returns an exact-version command only when the corresponding offline-signed agent release is cached and verified.
      responses:
        '200':
          description: Agent bootstrap availability and pinned connection details.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/AgentInstallCommand'}
        '400': {$ref: './components.yaml#/components/responses/BadRequest'}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
...
  /agents/{agent_id}/revoke:
    post:
      tags: [Agents]
      operationId: revokeAgent
      parameters:
        - {$ref: './components.yaml#/components/parameters/AgentID'}
        - {$ref: './components.yaml#/components/parameters/CSRFToken'}
        - {$ref: './components.yaml#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/RevokeAgentRequest'}
      responses:
        '200':
          description: Certificate revoked and active control stream closed.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/Agent'}
        '400': {$ref: './components.yaml#/components/responses/BadRequest'}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
        '403': {$ref: './components.yaml#/components/responses/ReauthenticationRequired'}
        '404': {$ref: './components.yaml#/components/responses/NotFound'}
        '409': {$ref: './components.yaml#/components/responses/Conflict'}
  /enrollment-requests:
    get:
      tags: [Enrollment]
      operationId: listEnrollmentRequests
      parameters:
        - {$ref: './components.yaml#/components/parameters/PageSize'}
        - {$ref: './components.yaml#/components/parameters/PageToken'}
        - {name: state, in: query, schema: {$ref: './components.yaml#/components/schemas/EnrollmentState'}}
      responses:
        '200':
          description: Enrollment requests visible to administrators.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/EnrollmentRequestList'}
    post:
      tags: [Enrollment]
      operationId: requestEnrollment
      description: Agent-facing endpoint. The manager parses the CSR and derives its fingerprint and verification code; the body never contains a private key.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/CreateEnrollmentRequest'}
      responses:
        '202':
          description: Request is awaiting approval or a pre-approved provision challenge.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/EnrollmentRequest'}
        '409': {$ref: './components.yaml#/components/responses/Conflict'}
        '422': {$ref: './components.yaml#/components/responses/ValidationFailed'}
        '429': {$ref: './components.yaml#/components/responses/RateLimited'}
...
  /enrollment-requests/{enrollment_id}/approve:
    post:
      tags: [Enrollment]
      operationId: approveEnrollmentRequest
      parameters:
        - {$ref: './components.yaml#/components/parameters/EnrollmentID'}
        - {$ref: './components.yaml#/components/parameters/CSRFToken'}
        - {$ref: './components.yaml#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/ApproveEnrollmentRequest'}
      responses:
        '200':
          description: Enrollment approved and identity assigned; the agent retrieves its certificate by proving possession of the CSR private key.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/EnrollmentRequest'}
        '403': {$ref: './components.yaml#/components/responses/ReauthenticationRequired'}
        '404': {$ref: './components.yaml#/components/responses/NotFound'}
        '409': {$ref: './components.yaml#/components/responses/Conflict'}
  /enrollment-requests/{enrollment_id}/status:
    post:
      tags: [Enrollment]
      operationId: getEnrollmentStatus
      description: |
        Agent-facing polling endpoint. Proof is an Ed25519 signature made by the
        CSR private key over `bored-manager enrollment status:{enrollment_id}`.
        This binds collection to the enrollment identity without a bearer token.
      security: []
      parameters:
        - {$ref: './components.yaml#/components/parameters/EnrollmentID'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/EnrollmentStatusProof'}
      responses:
        '200':
          description: Signed certificate chain and assigned identity after proof-of-possession verification.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/EnrollmentApproval'}
        '202':
          description: Request remains pending approval.
          headers:
            Retry-After: {schema: {type: integer, minimum: 1, default: 5}}
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/EnrollmentRequest'}
        '403': {$ref: './components.yaml#/components/responses/Forbidden'}
        '404': {$ref: './components.yaml#/components/responses/NotFound'}
        '409': {$ref: './components.yaml#/components/responses/Conflict'}
        '429': {$ref: './components.yaml#/components/responses/RateLimited'}
  /enrollment-requests/{enrollment_id}/reject:
    post:
      tags: [Enrollment]
      operationId: rejectEnrollmentRequest
      parameters:
        - {$ref: './components.yaml#/components/parameters/EnrollmentID'}
        - {$ref: './components.yaml#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/RejectEnrollmentRequest'}
      responses:
        '200':
          description: Rejected enrollment request.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/EnrollmentRequest'}
        '403': {$ref: './components.yaml#/components/responses/ReauthenticationRequired'}
        '404': {$ref: './components.yaml#/components/responses/NotFound'}
        '409': {$ref: './components.yaml#/components/responses/Conflict'}
...
  /docker-hosts:
    get:
      tags: [Docker Hosts]
      operationId: listDockerHosts
      parameters:
        - {$ref: './components.yaml#/components/parameters/PageSize'}
        - {$ref: './components.yaml#/components/parameters/PageToken'}
      responses:
        '200':
          description: Registered Docker hosts.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/DockerHostList'}
    post:
      tags: [Docker Hosts]
      operationId: createDockerHost
      parameters:
        - {$ref: './components.yaml#/components/parameters/CSRFToken'}
        - {$ref: './components.yaml#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: './components.yaml#/components/schemas/CreateDockerHostRequest'}
      responses:
        '201':
          description: Host saved after strict host-key preflight.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/DockerHost'}
        '409': {$ref: './components.yaml#/components/responses/Conflict'}
        '422': {$ref: './components.yaml#/components/responses/ValidationFailed'}
  /docker-hosts/local/discovery:
    get:
      tags: [Docker Hosts]
      operationId: discoverLocalDocker
      description: Probes `/var/run/docker.sock` directly without invoking the Docker CLI.
      responses:
        '200':
          description: Current local Docker Engine probe result.
          content:
            application/json:
              schema: {$ref: './components.yaml#/components/schemas/DockerProbe'}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
...
  /events:
    get:
      tags: [Events]
      operationId: connectEvents
      description: |
        Upgrades to WebSocket. Clients pass the last applied cursor and must
        refetch a REST snapshot when a `resync_required` envelope is received.
      x-websocket: true
      parameters:
        - name: cursor
          in: query
          schema: {$ref: './components.yaml#/components/schemas/EventCursor'}
        - name: Origin
          in: header
          required: true
          schema: {type: string, format: uri}
      responses:
        '101':
          description: WebSocket event stream; each text frame is an EventEnvelope.
          x-message-schema: {$ref: './components.yaml#/components/schemas/EventEnvelope'}
        '400': {$ref: './components.yaml#/components/responses/BadRequest'}
        '401': {$ref: './components.yaml#/components/responses/Unauthorized'}
        '403': {$ref: './components.yaml#/components/responses/Forbidden'}
  /artifacts/releases/{tag}/{name}:
    get:
      tags: [Bootstrap Artifacts]
      operationId: downloadVerifiedAgentArtifact
      description: Serves only a whitelisted artifact from the manager's matching offline-signed release cache.
      security: []
      parameters:
        - name: tag
          in: path
          required: true
          schema: {type: string, pattern: '^v[0-9]+\.[0-9]+\.[0-9]+(?:-[0-9A-Za-z.-]+)?$'}
        - name: name
          in: path
          required: true
          schema: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._+-]{0,127}$'}
      responses:
        '200':
          description: Verified immutable release artifact.
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        '404': {$ref: './components.yaml#/components/responses/NotFound'}
        '503': {$ref: './components.yaml#/components/responses/ServiceUnavailable'}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bm_session